Blob save working for encounters

+++ app/femr/business/services/system/PhotoService.java
@@ -45,6 +45,7 @@ import java.nio.charset.Charset;
 import java.nio.file.*;
 import java.util.ArrayList;
 import java.util.List;
+import java.io.ByteArrayOutputStream;
 
 public class PhotoService implements IPhotoService {
 
@@ -86,12 +87,24 @@ public class PhotoService implements IPhotoService {
             f.mkdirs();
     }
 
+    private byte[] convertBufferedImageToByteArray(BufferedImage img) {
+        if(img != null) {
+            try {
+                ByteArrayOutputStream baos = new ByteArrayOutputStream();
+                ImageIO.write(img, "jpg", baos);
+            } catch (Exception ex) {
+            }
+        }
+        return null;
+    }
+
     /**
      * {@inheritDoc}
      */
     @Override
     public ServiceResponse<Boolean> createPatientPhoto(String imageString, int patientId, Boolean deleteFlag) {
         ServiceResponse<Boolean> response = new ServiceResponse<>();
+        byte[] photoData;
 
         try {
             IPatient patient = patientRepository.retrievePatientById(patientId);
@@ -114,7 +127,9 @@ public class PhotoService implements IPhotoService {
 
                 if (patient.getPhoto() == null) {
                     //Create new photo Id record, pass in binary photo data if _bUseDbPhotoStorage is true
-                    IPhoto pPhoto = photoRepository.createPhoto("", imageFileName, _bUseDbPhotoStorage ? bufferedImage : null);
+                    IPhoto pPhoto = photoRepository.createPhoto("",
+                            imageFileName,
+                            _bUseDbPhotoStorage ? convertBufferedImageToByteArray(bufferedImage) : null);
                     patient.setPhoto(pPhoto);
                     patientRepository.savePatient(patient);
                 } else {
@@ -144,7 +159,35 @@ public class PhotoService implements IPhotoService {
     }
 
     @Override
-    public ServiceResponse<byte[]> retrievePhotoData(int patientId) {
+    public ServiceResponse<byte[]> retrievePhotoData(int photoId) {
+        ServiceResponse<byte[]> response = new ServiceResponse<>();
+        try {
+            IPhoto photo = photoRepository.retrievePhotoById(photoId);
+            if(photo != null) {
+                if(!_bUseDbPhotoStorage) {
+                    //Read data from file and return:
+                    File encPhoto = new File(_encounterPhotoPath + photo.getFilePath());
+                    if(encPhoto.canRead()) {
+                        byte[] photoData = Files.readAllBytes(Paths.get(_encounterPhotoPath + photo.getFilePath()));
+                        response.setResponseObject(photoData);
+                    }
+                } else {
+                    //Data is already loaded in Photo record, let's return it!
+                    response.setResponseObject(photo.getPhotoBlob());
+                }
+            }
+
+        }
+        catch (Exception ex) {
+            response.addError("", ex.getMessage());
+            return response;
+        }
+
+        return response;
+    }
+
+    @Override
+    public ServiceResponse<byte[]> retrievePatientPhotoData(int patientId) {
         ServiceResponse<byte[]> response = new ServiceResponse<>();
         try {
             IPatient patient = patientRepository.retrievePatientById(patientId);
@@ -167,6 +210,8 @@ public class PhotoService implements IPhotoService {
         }
         return response;
     }
+
+
 
     /**
      * {@inheritDoc}
@@ -276,16 +321,18 @@ public class PhotoService implements IPhotoService {
 
     private void saveNewEncounterImage(File image, PatientEncounterItem patientEncounter, String descriptionText) {
         try {
-            String imageFileName;
+            String imageFileName = "N/A";
 
             //Create photo record:
             IPhoto pPhoto = photoRepository.createPhoto(descriptionText, "", null);
             //this is redundant...?
             IPhoto editPhoto = photoRepository.retrievePhotoById(pPhoto.getId());
 
-            //Build the filepath for the image
-            imageFileName = "Patient_" + patientEncounter.getPatientItem().getId()
-                    + "_Enc_" + patientEncounter.getId() + "_Photo_" + editPhoto.getId();
+            if(!_bUseDbPhotoStorage) {
+                //Build the filepath for the image
+                imageFileName = "Patient_" + patientEncounter.getPatientId()
+                        + "_Enc_" + patientEncounter.getId() + "_Photo_" + editPhoto.getId();
+            }
 
             //Since the photo ID is part of the file name
             //  I am setting the filePath field after the record is created
@@ -294,7 +341,16 @@ public class PhotoService implements IPhotoService {
             //Link photo record in photoEncounter table
             photoRepository.createEncounterPhoto(editPhoto.getId(), patientEncounter.getId());
 
-            createPhotoOnFilesystem(image, this._encounterPhotoPath + imageFileName);
+            if(!_bUseDbPhotoStorage) {
+                createPhotoOnFilesystem(image, this._encounterPhotoPath + imageFileName);
+            } else {
+                //Read image data, update blob field:
+                byte[] imgData = Files.readAllBytes(Paths.get(image.getAbsolutePath()));
+                photoRepository.updatePhotoData(pPhoto.getId(), imgData);
+            }
+
+
+
 
         } catch (Exception ex) {
 
@@ -317,7 +373,8 @@ public class PhotoService implements IPhotoService {
 
             photoRepository.deleteEncounterPhotosByPhotoId(savedPhoto.getId());
 
-            photoRepository.deletePhotoFromFilesystemById(filePath + savedPhoto.getFilePath());
+            if(!_bUseDbPhotoStorage)
+                photoRepository.deletePhotoFromFilesystemById(filePath + savedPhoto.getFilePath());
 
             photoRepository.deletePhotoById(savedPhoto.getId());
         }
